Extract Assignment and Highlight layers from zone layers
