+++ app/src/app/components/Map/VtdBlockLayers.tsx
@@ -4,8 +4,8 @@ import {useMapStore} from '@/app/store/mapStore';
 import {useMapControlsStore} from '@/app/store/mapControlsStore';
 import {useLayoutEffect} from 'react';
 import {useMap} from 'react-map-gl/maplibre';
-import ZoneChildLayerGroup from './ZoneChildLayerGroup';
-import ZoneParentLayerGroup from './ZoneParentLayerGroup';
+import ZoneChildLayerGroup from './ZoneLayers/ZoneChildLayerGroup';
+import ZoneParentLayerGroup from './ZoneLayers/ZoneParentLayerGroup';
 import {DemographicLayer} from './DemographicLayer';
 import {HoverLayerGroup} from './HoverLayerGroup';
 import {demographyCache} from '@/app/utils/demography/demographyCache';
